refactor: address PR review feedback - separate debug concerns and improve maintainability
✅ **PR Review Responses:**

🔄 **Separated Debug and Export Concerns:**
- Created dedicated `src/store/atoms/debug.ts` for debug-related atoms
- Created `src/store/hooks/useDebug.ts` for debug state management hooks
- Removed debug functionality from export-related files for better modularity
- Updated import paths throughout the application

📦 **Improved Code Organization:**
- Provider import alias already implemented (JotaiRootProvider)
- SEARCH_SCOPES constant extraction already completed
- Underscore prefix issues in App.tsx already resolved

🎯 **Code Quality Improvements:**
- Better separation of concerns between debug, export, and other state domains
- Cleaner module boundaries with focused responsibilities
- Maintained all existing functionality while improving architecture

✅ **Quality Assurance:**
- All 455+ tests passing
- TypeScript strict mode compliance maintained
- No breaking changes to existing functionality
- Improved maintainability and code organization

This refactoring addresses CodeRabbit and Gemini Code Assist feedback while maintaining the high-quality atomic state management architecture.
